requirements.txt
pycryptodomex==3.12.0
python_bcrypt==0.3.2
